homework5/requirements.txt
Pillow
future
matplotlib
numpy
tensorboard>=1.14
torch>=1.6.0
torchvision>=0.7.0
PySuperTuxKart
